requirements.txt
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
idna==3.6
jmespath==1.0.1
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
urllib3==2.0.7
